Move question re-ordering logic server side
Begin to move the logic for re-orering questions server side.  Having
all of the re-numbering done in the browser is far too slow and
unreliable to be done in the browser.

This implementation only effects the minimum number of questions needed,
by using where the question moved from and where it moved to.  If a
question moves by a large amount, there still is a noticable lag.  If
this is a problem, we can consider moving this to a background process.

Still left to do:
* Fix manual drag and drop of questions to work correctly
* Fix add question to renumber everything that comes after it correctly
* Fix delete question to renumber everything that came after it
* correctly